requirements.txt
opencv-python-headless
gradio
transformers
torch
accelerate
numpy
pillow
scikit-learn